+++ pkg/protocol/http1/req/header_test.go
@@ -49,7 +49,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/cloudwego/hertz/internal/bytesconv"
 	"github.com/cloudwego/hertz/pkg/common/test/assert"
 	"github.com/cloudwego/hertz/pkg/common/test/mock"
 	"github.com/cloudwego/hertz/pkg/protocol"
@@ -248,18 +247,6 @@ func TestRequestHeaderCookie(t *testing.T) {
 	if len(h.Cookie("привет")) > 0 {
 		t.Fatalf("Unexpected cookie found: %q", h.Cookie("привет"))
 	}
-}
-
-func TestRequestHeaderCookies(t *testing.T) {
-	var h protocol.RequestHeader
-	h.SetCookie("foo", "bar")
-	h.SetCookie("привет", "мир")
-	cookies := h.Cookies()
-	assert.DeepEqual(t, 2, len(cookies))
-	assert.DeepEqual(t, "foo", bytesconv.B2s(cookies[0].Key()))
-	assert.DeepEqual(t, "bar", bytesconv.B2s(cookies[0].Value()))
-	assert.DeepEqual(t, "привет", bytesconv.B2s(cookies[1].Key()))
-	assert.DeepEqual(t, "мир", bytesconv.B2s(cookies[1].Value()))
 }
 
 func TestRequestRawHeaders(t *testing.T) {
